[DA] add code_version_changed AutomationCondition
