+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/asset_condition_tests/test_code_version_changed_condition.py
@@ -1,0 +1,41 @@
+from dagster import AutomationCondition
+
+from ..scenario_specs import one_asset
+from .asset_condition_scenario import AutomationConditionScenarioState
+
+
+def test_code_version_changed_condition() -> None:
+    state = AutomationConditionScenarioState(
+        one_asset, automation_condition=AutomationCondition.code_version_changed()
+    ).with_asset_properties(code_version="1")
+
+    # not changed
+    state, result = state.evaluate("A")
+    assert result.true_subset.size == 0
+
+    # still not changed
+    state, result = state.evaluate("A")
+    assert result.true_subset.size == 0
+
+    # newly changed
+    state = state.with_asset_properties(code_version="2")
+    state, result = state.evaluate("A")
+    assert result.true_subset.size == 1
+
+    # not newly changed
+    state, result = state.evaluate("A")
+    assert result.true_subset.size == 0
+
+    # newly changed
+    state = state.with_asset_properties(code_version="3")
+    state, result = state.evaluate("A")
+    assert result.true_subset.size == 1
+
+    # newly changed
+    state = state.with_asset_properties(code_version="2")
+    state, result = state.evaluate("A")
+    assert result.true_subset.size == 1
+
+    # not newly changed
+    state, result = state.evaluate("A")
+    assert result.true_subset.size == 0
